Moved multithread, added unique result filenames
